# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go e-commerce backend with Docker, Elasticsearch, Redis, RabbitMQ.",
        "title": "E-Commerce API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Hasan Can Sevim",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/products/search": {
            "get": {
                "description": "Smart search on products using Elasticsearch (Fuzzy \u0026 Wildcard).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Search Products",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Aranacak Kelime (Örn: 'laptop')",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ProductResponse": {
            "type": "object",
            "properties": {
                "base_price": {
                    "type": "number"
                },
                "category_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_featured": {
                    "type": "boolean"
                },
                "meta_description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "slug": {
                    "type": "string"
                },
                "stock_quantity": {
                    "type": "integer"
                },
                "store_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}